♻️ Refactor: webrtc 파일들을 data 레이어로 migration
#17
