+++ app/src/main/java/com/example/saybettereducator/data/model/FirebaseFieldNames.kt
@@ -1,4 +1,4 @@
-package com.example.saybettereducator.utils
+package com.example.saybettereducator.data.model
 
 object FirebaseFieldNames {
     const val STATUS = "status"
