+++ app/src/main/java/com/example/saybettereducator/data/model/MainServiceActions.kt
@@ -1,4 +1,4 @@
-package com.example.saybettereducator.utils.webrtc.service
+package com.example.saybettereducator.data.model
 
 enum class MainServiceActions {
     START_SERVICE, SETUP_VIEWS, END_CALL, SWITCH_CAMERA, TOGGLE_AUDIO, TOGGLE_VIDEO;
