+++ app/src/main/java/com/example/saybettereducator/utils/webrtcObserver/MySdpObserver.kt
@@ -1,4 +1,4 @@
-package com.example.saybettereducator.utils.webrtc.webrtcClient
+package com.example.saybettereducator.utils.webrtcObserver
 
 import org.webrtc.SdpObserver
 import org.webrtc.SessionDescription
